fix luci-app-adguardhome

+++ applications/luci-app-adguardhome/luasrc/controller/AdGuardHome.lua
@@ -1,30 +1,47 @@
-module("luci.controller.AdGuardHome", package.seeall)
-local fs = require "nixio.fs"
-local http = require "luci.http"
-local uci = require"luci.model.uci".cursor()
+module("luci.controller.AdGuardHome",package.seeall)
+local fs=require"nixio.fs"
+local http=require"luci.http"
+local uci=require"luci.model.uci".cursor()
 function index()
-local page = entry({"admin", "services", "AdGuardHome"},alias("admin", "services", "AdGuardHome", "base"),_("AdGuard Home"))
-page.order = 11
-page.dependent = true
-page.acl_depends = { "luci-app-adguardhome" }
-    entry({"admin", "services", "AdGuardHome", "base"}, cbi("AdGuardHome/base"),  _("Base Setting"), 1).leaf = true
-    entry({"admin", "services", "AdGuardHome", "log"}, form("AdGuardHome/log"), _("Log"), 2).leaf = true
-    entry({"admin", "services", "AdGuardHome", "manual"}, cbi("AdGuardHome/manual"), _("Manual Config"), 3).leaf = true
-    entry({"admin", "services", "AdGuardHome", "status"}, call("act_status")).leaf = true
-    entry({"admin", "services", "AdGuardHome", "check"}, call("check_update"))
-    entry({"admin", "services", "AdGuardHome", "doupdate"}, call("do_update"))
-    entry({"admin", "services", "AdGuardHome", "getlog"}, call("get_log"))
-    entry({"admin", "services", "AdGuardHome", "dodellog"}, call("do_dellog"))
-    entry({"admin", "services", "AdGuardHome", "reloadconfig"}, call("reload_config"))
-    entry({"admin", "services", "AdGuardHome", "gettemplateconfig"}, call("get_template_config"))
-end
+entry({"admin", "services", "AdGuardHome"},alias("admin", "services", "AdGuardHome", "base"),_("AdGuard Home"), 10).dependent = true
+entry({"admin","services","AdGuardHome","base"},cbi("AdGuardHome/base"),_("Base Setting"),1).leaf = true
+entry({"admin","services","AdGuardHome","log"},form("AdGuardHome/log"),_("Log"),2).leaf = true
+entry({"admin","services","AdGuardHome","manual"},cbi("AdGuardHome/manual"),_("Manual Config"),3).leaf = true
+entry({"admin","services","AdGuardHome","status"},call("act_status")).leaf=true
+entry({"admin", "services", "AdGuardHome", "check"}, call("check_update"))
+entry({"admin", "services", "AdGuardHome", "doupdate"}, call("do_update"))
+entry({"admin", "services", "AdGuardHome", "getlog"}, call("get_log"))
+entry({"admin", "services", "AdGuardHome", "dodellog"}, call("do_dellog"))
+entry({"admin", "services", "AdGuardHome", "reloadconfig"}, call("reload_config"))
+entry({"admin", "services", "AdGuardHome", "gettemplateconfig"}, call("get_template_config"))
+end 
 function get_template_config()
 	local b
 	local d=""
-	for cnt in io.lines("/tmp/resolv.conf.d/resolv.conf.auto") do
-		b=string.match (cnt,"^[^#]*nameserver%s+([^%s]+)$")
-		if (b~=nil) then
-			d=d.."  - "..b.."\n"
+	local rcauto=uci:get("dhcp","@dnsmasq[0]","resolvfile")
+	if (rcauto == nil) then
+		for fle in fs.dir("/var/etc") do
+			if fle ~="." and fle ~=".."then
+				tf="/var/etc/"..fle
+				if string.match(tf,"/var/etc/dnsmasq.conf.") then
+					if tf and fs.access(tf) then
+						for le in io.lines(tf) do
+							sf=string.match (le,"^resolv%-file=(%S+)")
+								if (sf ~=nil) then
+								rcauto=sf
+							end
+						end
+					end
+				end
+			end
+		end
+	end
+	if rcauto and fs.access(rcauto) then
+		for cnt in io.lines(rcauto) do
+			b=string.match (cnt,"^[^#]*nameserver%s+([^%s]+)$")
+			if (b~=nil) then
+				d=d.."    - "..b.."\n"
+			end
 		end
 	end
 	local f=io.open("/usr/share/AdGuardHome/AdGuardHome_template.yaml", "r+")
@@ -68,10 +85,10 @@ function do_update()
 	else
 		arg=""
 	end
-	if arg=="force" then
+	if fs.access("/var/run/update_core") then
+		if arg=="force" then
 			luci.sys.exec("kill $(pgrep /usr/share/AdGuardHome/update_core.sh) ; sh /usr/share/AdGuardHome/update_core.sh "..arg.." >/tmp/AdGuardHome_update.log 2>&1 &")
-	
-
+		end
 	else
 		luci.sys.exec("sh /usr/share/AdGuardHome/update_core.sh "..arg.." >/tmp/AdGuardHome_update.log 2>&1 &")
 	end
